phase3/backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-socketio>=5.10.0
pydantic>=2.5.0
pytest>=7.4.3
pytest-asyncio>=0.21.1
python-multipart==0.0.6
aiofiles==23.2.1
psutil==5.9.8
httpx>=0.27.0
pytest-cov>=4.1.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
PyJWT[crypto]>=2.8.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
bcrypt>=4.0.0
python-socketio[asyncio_client]>=5.10.0
web3>=6.0.0
slowapi>=0.1.9
fastapi-csrf-protect>=0.3.0
itsdangerous>=2.1.0
cryptography>=41.0.0
alembic>=1.13.0
langgraph>=0.0.20
langchain>=0.1.0
langchain-core>=0.1.0
docker>=6.0.0
redis>=4.5.0
pandas>=2.0.0
numpy>=1.24.0
jsonschema>=4.17.0
qrcode[pil]>=7.4.2
Pillow>=10.0.0
prometheus-client>=0.19.0
asteval>=0.9.31
eth-account>=0.10.0
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0
chromadb>=0.4.0
pgvector>=0.2.0
reportlab>=4.0.0
python-telegram-bot>=20.0
aiohttp>=3.9.0
mnemonic>=0.20
bip32utils>=0.3.post4
APScheduler>=3.10.0
ddgs>=1.0.0
